% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.in.R
\docType{data}
\name{df.in}
\alias{df.in}
\title{Input example dataset}
\format{A matrix with 50 rows and 10 variables, each row is an individual
subject}
\usage{
df.in
}
\description{
A dataset containing example longitudinal data
}
\keyword{datasets}

