% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.in.R
\docType{data}
\name{df2.in}
\alias{df2.in}
\title{Input example dataset}
\format{A matrix with 400000 rows and 50 variables, each row is an individual
subject}
\usage{
df2.in
}
\description{
A dataset containing example longitudinal data
}
\keyword{datasets}

