% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dfviewr.R
\name{dfviewr}
\alias{dfviewr}
\title{data frame viewer}
\usage{
dfviewr(df, col = c("red", "blue", "green", "white", "black"), axes = FALSE,
  main = "", main.adj = 0, cex.axis = 1.75, gridlines = TRUE,
  legend = TRUE, ...)
}
\arguments{
\item{df}{a dataframe}

\item{col}{a vector of 5 colors.  The first is for dates, second for factors, third for characters, fourth for numeric, and fifth for NA.}

\item{gridlines}{(logical) default TRUE.}

\item{legend}{(logical) defaults to TRUE.}

\item{...}{Additional stuff to be passed to \code{image}}
}
\description{
data frame viewer
}

