% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er.R
\name{er}
\alias{er}
\title{entire-row sorting}
\usage{
er(X, orderVar = min(X, na.rm = TRUE), naLast = FALSE)
}
\arguments{
\item{X}{a matrix: each row a subject, each column a time/location across subjects}

\item{orderVar}{a scalar; one of the values in X upon which the percentage of occurence will determine the ordering of the rows. 
Defaults to the minimum value of X}

\item{naLast}{defaults to false; should NA values be sorted to first position or last. Defaults to FALSE.}
}
\description{
Global sorting of rows
}

