% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/lasagna_plain.R
\name{lasagna_plain}
\alias{lasagna_plain}
\title{Lasagna Function: no axes, no title, no legend, no nothing}
\usage{
lasagna_plain(X, col = rainbow_hcl(100), axes = FALSE, ...)
}
\arguments{
\item{X}{a matrix: each row a subject, each column a time/location across subjects}

\item{col}{colors.  See "Escaping RGB Land" and the vignette for the colorspace package
for followable advice on colors.}

\item{axes}{(logical) axes on or off?}

\item{...}{Additional stuff to be passed to \code{image}}
}
\description{
Lasagna Function: no axes, no title, no legend, no nothing
}

