% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc.cont.R
\name{wc.cont}
\alias{wc.cont}
\title{Within-column sorting of continuous values}
\usage{
wc.cont(X, naLast = FALSE)
}
\arguments{
\item{X}{a matrix: each row a subject, each column a time/location across subjects}

\item{naLast}{defaults to false; should NA values be sorted to first position or last. Defaults to FALSE.}

\item{colorSeq}{a vector with the unique elements of X presented in the order of the desired
resultant sorting}
}
\description{
Sorts continuous matrices within columns
}
\details{
\code{wc} sorts within-columns, scrambling the subject-specific nature 
of the rows but revealing a group-level temporal patterns.  As with \code{wr} 
there is a discrete and continuous implementation \code{wc.disc} 
and \code{wc.cont}
}

