% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfviewr.R
\name{dfviewr}
\alias{dfviewr}
\title{data frame viewer}
\usage{
dfviewr(df, col = c("red", "blue", "green", "white", "black"), axes = FALSE,
  main = "", gridlines = TRUE, legend = TRUE, ...)
}
\arguments{
\item{df}{a dataframe}

\item{col}{a vector of 5 colors.  The first is for dates, 
second for factors, third for characters, fourth for numeric, and fifth for NA.}

\item{axes}{Axes passed to \code{lasagna}}

\item{main}{Title passed to \code{lasagna}}

\item{gridlines}{(logical) default TRUE,  passed to \code{lasagna}}

\item{legend}{(logical) defaults to TRUE. Creates a legend}

\item{...}{Additional argument to be passed to \code{image}}
}
\description{
Views data.frames with a legend
}

