% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dfviewr.R
\name{dfviewr}
\alias{dfviewr}
\title{data frame viewer}
\usage{
dfviewr(df, col = c("red", "blue", "green", "white", "black"), axes = FALSE,
  main = "", main.adj = 0, cex.axis = 1.75, gridlines = TRUE,
  legend = TRUE, ...)
}
\arguments{
\item{df}{a dataframe}

\item{axes}{(logical) TRUE invokes the default axes from image().  FALSE invokes best guess (to be better defined later)}

\item{main}{as in plot(), the text of the title.  Defaults to "(A)  Initial Lasagna Plot"}

\item{main.adj}{the value for `adj` in title().  Defaults to 0, which is left-aligned.}

\item{cex.axis}{the cex.axis value if `axes` is FALSE.}

\item{gridlines}{(logical) default TRUE.}

\item{legend}{(logical) defaults to FALSE.  TRUE invokes image.plot() from fields package instead of image() and thus displays a legend.}

\item{...}{Additional stuff to be passed to \code{image}}
}
\description{
data frame viewer
}

