% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/er.R
\name{er}
\alias{er}
\title{entire-row sorting}
\usage{
er(X, orderVar = c(), naLast = F)
}
\arguments{
\item{X}{a matrix: each row a subject, each column a time/location across subjects}

\item{orderVar}{a vector with the unique elements of X presented in the order of the desired
resultant sorting}

\item{naLast}{defaults to false; should NA values be sorted to first position or last. Defaults to FALSE.}
}
\description{
entire-row sorting
}

