% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/lasagna.R
\name{lasagna}
\alias{lasagna}
\title{Lasagna Function}
\usage{
lasagna(X, col = rainbow_hcl(100), axes = FALSE, ...)
}
\arguments{
\item{X}{a matrix: each row a subject, each column a time/location across subjects}

\item{col}{colors}

\item{axes}{(logical) axes on or off?}

\item{...}{Additional stuff to be passed to \code{image}}
}
\description{
Lasagna Function
}

