% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasagna.R
\name{lasagna}
\alias{lasagna}
\title{Lasagna Function}
\usage{
lasagna(X, col = rainbow_hcl(length(unique(c(X)))), image.axes = FALSE,
  xaxis = TRUE, yaxis = TRUE, main = "(A)  Initial Lasagna Plot",
  main.adj = 0, cex.axis = 1.75, gridlines = FALSE, legend = FALSE, ...)
}
\arguments{
\item{X}{a matrix: each row a subject, each column a time/location across subjects; preferably rownames are the ids.}

\item{col}{colors.  See "Escaping RGB Land" and the vignette for the colorspace package
for followable advice on colors.}

\item{xaxis}{Use best guess x-axis if \code{image.axis = FALSE}}

\item{yaxis}{Use best guess y-axis if \code{image.axis = FALSE}}

\item{main}{as in plot(), the text of the title.  Defaults to "(A)  Initial Lasagna Plot"}

\item{main.adj}{the value for `adj` in title().  Defaults to 0, which is left-aligned.}

\item{cex.axis}{the cex.axis value if `axes` is FALSE.}

\item{gridlines}{(logical) Add gridlines to the plot.  Useful if image is not
too big.}

\item{legend}{(logical) defaults to FALSE.  TRUE invokes image.plot() from fields package instead of image() and thus displays a legend.}

\item{...}{Additional stuff to be passed to \code{image}}

\item{axes}{(logical) TRUE invokes the default axes from image().
FALSE invokes best guess, and better control using \code{xaxis/yaxis}.}
}
\description{
Generic Lasagna Function
}

