% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/wc.disc.R
\name{wc.disc}
\alias{wc.disc}
\title{within-column sorting of discrete values}
\usage{
wc.disc(X, colorSeq = rev(sort(unique(X))), naLast = F)
}
\arguments{
\item{X}{a matrix: each row a subject, each column a time/location across subjects}

\item{colorSeq}{a vector with the unique elements of X presented in the order of the desired
resultant sorting.  Defaults to having lowest value sort to the bottom, highest value to the top.}

\item{naLast}{defaults to false; should NA values be sorted to first position or last. Defaults to FALSE.
## within-column for discrete}
}
\description{
within-column sorting of discrete values
}

