% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/wr.disc.R
\name{wr.disc}
\alias{wr.disc}
\title{within-row sorting of discrete values}
\usage{
wr.disc(X, colorSeq = sort(unique(X)), naLast = F)
}
\arguments{
\item{X}{a matrix: each row a subject, each column a time/location across subjects}

\item{colorSeq}{a vector with the unique elements of X presented in the order of the desired
resultant sorting}

\item{naLast}{defaults to false; should NA values be sorted to first position or last. Defaults to FALSE.
## within-row for discrete}
}
\description{
within-row sorting of discrete values
}

